Fix possible StackOverflow in WorldGenAdvLakes
Really bad idea to check the world state 8 blocks away from current position.  That location could be in another chunk, which might need generating, which might call this code, which might.... oh my goodness...!

I can see this is how it is done by Mojang for their lakes but don't think it is a good idea.  I believe this is giving me a StackOverflow.